+++ py/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-
+- Support comparing release versions without build code. ([#5376](https://github.com/getsentry/relay/pull/5376))
 
 ## 0.9.21